receitas/python/blinkist_project/requirements.txt
selenium
pytest-appium
pytest
Appium-Python-Client